refactor: rename span_task to span
