+++ spacy_llm/tasks/util/span.py
